requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-decouple==3.7
pytz==2022.7.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
